Some more format fixes that got away.

+++ hc_extract_sh_layer.c
@@ -113,7 +113,7 @@ int main(int argc, char **argv)
     if(mode != 4){
       /* SH header */
       if(short_format && loop)
-	fprintf(stdout,"%g\n",HC_Z_DEPTH(model->r[ilayer]));
+	fprintf(stdout,"%g\n",(double)HC_Z_DEPTH(model->r[ilayer]));
       sh_print_parameters_to_file((sol+ilayer*shps_read),shps,
 				  ilayer,nset,(HC_PREC)(HC_Z_DEPTH(model->r[ilayer])),
 				  stdout,short_format,FALSE,verbose);
@@ -123,38 +123,38 @@ int main(int argc, char **argv)
       /*  */
       if(verbose)
 	fprintf(stderr,"%s: printing x_r SHE at layer %i (depth: %g)\n",
-		argv[0],ilayer,HC_Z_DEPTH(model->r[ilayer]));
+		argv[0],ilayer,(double)HC_Z_DEPTH(model->r[ilayer]));
       sh_print_coefficients_to_file((sol+ilayer*shps_read),shps,stdout,fac,FALSE,verbose);
       break;
     case 2:
       /*  */
       if(verbose)
 	fprintf(stderr,"%s: printing x_pol x_tor SHE at layer %i (depth: %g)\n",
-		argv[0],ilayer,HC_Z_DEPTH(model->r[ilayer]));
+		argv[0],ilayer,(double)HC_Z_DEPTH(model->r[ilayer]));
       sh_print_coefficients_to_file((sol+ilayer*shps_read+1),shps,stdout,fac,FALSE,verbose);
       break;
     case 3:
       /* mode == 3 */
       if(verbose)
 	fprintf(stderr,"%s: printing x_r x_pol x_tor SHE at layer %i (depth: %g)\n",
-		argv[0],ilayer,HC_Z_DEPTH(model->r[ilayer]));
+		argv[0],ilayer,(double)HC_Z_DEPTH(model->r[ilayer]));
       sh_print_coefficients_to_file((sol+ilayer*shps_read),shps,stdout,fac,FALSE,verbose);
       break;
     case 4:
-      fprintf(stdout,"%5i %11g\n",ilayer,HC_Z_DEPTH(model->r[ilayer]));
+      fprintf(stdout,"%5i %11g\n",ilayer,(double)HC_Z_DEPTH(model->r[ilayer]));
       break;
     case 5:
       /*  */
       if(verbose)
 	fprintf(stderr,"%s: printing x_pol SHE at layer %i (depth: %g)\n",
-		argv[0],ilayer,HC_Z_DEPTH(model->r[ilayer]));
+		argv[0],ilayer,(double)HC_Z_DEPTH(model->r[ilayer]));
       sh_print_coefficients_to_file((sol+ilayer*shps_read+1),shps,stdout,fac,FALSE,verbose);
       break;
     case 6:
       /*  */
       if(verbose)
 	fprintf(stderr,"%s: printing x_tor SHE at layer %i (depth: %g)\n",
-		argv[0],ilayer,HC_Z_DEPTH(model->r[ilayer]));
+		argv[0],ilayer,(double)HC_Z_DEPTH(model->r[ilayer]));
       sh_print_coefficients_to_file((sol+ilayer*shps_read+2),shps,stdout,fac,FALSE,verbose);
       break;
